Merge rama_Alcocer3 - Pedidos, Verif Cédula, PeDetalles

+++ database/migrations/2024_07_03_013614_create_ingresos_table.php
@@ -15,8 +15,7 @@ return new class extends Migration
             $table->id();
             $table->date('fecha_ing');
             $table->string('descripcion_ing',150);
-            $table->integer('cantidad_ing');
-            
+            $table->decimal('cantidad_ing',8,2);
             $table->timestamps();
         });
     }
